Rename package error to errors, error Struct to Error
