+++ modules/ballerina-core/src/main/resources/MessagesBundle_en_US.properties
@@ -68,7 +68,7 @@ undefined.package.name = undefined package name ''{0}'' in ''{1}''
 template.expression.not.allowed.here = xml/json template expression is not allowed here
 connector.init.not.allowed.here = connector initializer is not allowed here
 only.count.1.allowed.this.version = only count 1 is allowed in this version
-only.error.type.here = only a struct type structurally equivalent to 'ballerina.lang.error:error' is allowed here
+only.error.type.here = only a struct type structurally equivalent to 'ballerina.lang.errors:Error' is allowed here
 break.stmt.not.allowed.here = break statement is not allowed here
 undefined.type.mapper = undefine type mapper ''{0}''
 incompatible.types.unknown.found = incompatible types: expected a ''{0}''
