+++ modules/ballerina-native/src/test/resources/lang/errors/duplicate-var-try-catch.bal
@@ -1,10 +1,10 @@
-import ballerina.lang.error;
+import ballerina.lang.errors;
 
 function testInvalid (){
     try{
         int a = 10;
-    } catch (error:error e){
-        error:error e = { msg: "test"};
+    } catch (errors:Error e){
+        errors:Error e = { msg: "test"};
         int b = 10;
     }
 }
